# Conflicts:
#	data.h2.mv.db
#	data.h2.trace.db
#	frontend/src/components/ProductsView/Product.js
#	frontend/src/components/ProductsView/ProductDetails.js
#	frontend/src/components/shopkeeper/ShopkeeperEditProduct.js
#	frontend/src/components/shopkeeper/ShopkeeperForm.js

+++ backend/src/main/java/net/yorksolutions/backend/controller/OrderController.java
@@ -1,6 +1,7 @@
 package net.yorksolutions.backend.controller;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import net.yorksolutions.backend.model.CartItem;
 import net.yorksolutions.backend.model.OrderDetails;
 import net.yorksolutions.backend.model.OrderItem;
 import net.yorksolutions.backend.model.Product;
@@ -12,6 +13,8 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Stream;
 
 @RestController
 @RequestMapping("/api/order")
@@ -28,13 +31,28 @@ public class OrderController {
     @Autowired
     CartItemRepository cartItemRepo;
 
+    @Autowired
+    ProductRepository productRepo;
+
+    @Autowired
+    CouponRepository couponRepo;
+
+
     class OrderItemsOutput extends Product{
         @JsonProperty
         private int quantity;
 
-        public OrderItemsOutput(Long id, String productName, String productDescription, String brand, Float unitPrice, int unitsInStock, String size, String color, Date productAvailable, Boolean discontinued, Boolean discountAvailable, String picture, Date dateReceived, int unitsReceived, int quantity) {
+        @JsonProperty
+        private Float regularPrice;
+
+        @JsonProperty
+        private Float salePrice;
+
+        public OrderItemsOutput(Long id, String productName, String productDescription, String brand, Float unitPrice, int unitsInStock, String size, String color, Date productAvailable, Boolean discontinued, Boolean discountAvailable, String picture, Date dateReceived, int unitsReceived, int quantity, Float regularPrice, Float salePrice) {
             super(id, productName, productDescription, brand, unitPrice, unitsInStock, size, color, productAvailable, discontinued, discountAvailable, picture, dateReceived, unitsReceived);
             this.quantity = quantity;
+            this.regularPrice = regularPrice;
+            this.salePrice = salePrice;
         }
     }
 
@@ -52,23 +70,32 @@ public class OrderController {
 
     @CrossOrigin
     @PostMapping("/add")
-    String createOrder (@RequestHeader Long cartId) {
+    String createOrder (@RequestHeader Long cartId, @RequestHeader Optional<String> couponCode) {
         var cart = cartRepo.findById(cartId).get();
         var userId = cart.userId;
         var total = cart.totalCost;
-        var order = new OrderDetails(userId, total);
-        orderDetailsRepo.save(order);
         var cartItems= cartItemRepo.findAllByCartId(cartId).orElseThrow();
-        var orderItems = new ArrayList<OrderItem>();
-        for(var item : cartItems){
-            var orderItem = new OrderItem(item.getProductId(), order.orderDetailsId, item.getQuantity());
-            orderItems.add(orderItem);
-        }
-        orderItemsRepo.saveAll(orderItems);
+        var order = new OrderDetails(userId, total);
+        var orderStatus = createOrder(cartItems, order, couponCode);
+        if (!orderStatus.equals("success"))
+            return orderStatus;
         cartItemRepo.deleteAllByCartId(cartId);
         cart.totalCost = 0F;
         cartRepo.save(cart);
 
+        return "success";
+    }
+
+
+
+    @CrossOrigin
+    @PostMapping("/addGuestOrder")
+    String createGuestOrder (@RequestBody List<CartItem> cartItems, @RequestHeader String email,
+                             @RequestHeader Float total, @RequestHeader Optional<String> couponCode) {
+        var order = new OrderDetails(email, total);
+        var orderStatus = createOrder(cartItems, order, couponCode);
+        if (!orderStatus.equals("success"))
+            return orderStatus;
         return "success";
     }
 
@@ -82,11 +109,32 @@ public class OrderController {
             Product p = (Product) itemDetail[0];
             OrderItem oi = (OrderItem) itemDetail[1];
             var orderDetails = new OrderItemsOutput(p.id, p.productName, p.productDescription, p.brand, p.unitPrice, p.unitsInStock, p.size,
-                    p.color, p.productAvailable, p.discontinued, p.discountAvailable, p.picture, p.dateReceived, p.unitsReceived, oi.getQuantity());
+                    p.color, p.productAvailable, p.discontinued, p.discountAvailable, p.picture, p.dateReceived, p.unitsReceived, oi.getQuantity(), oi.getRegularPrice(), oi.getSalePrice());
             orderInfo.add(orderDetails);
         }
 
         return new OrderHistoryOutput(orderInfo, order);
+    }
+
+    @CrossOrigin
+    @GetMapping("/all")
+    Iterable<OrderHistoryOutput> viewAllOrders(){
+        var orders = orderDetailsRepo.findAll();
+        List<OrderHistoryOutput> orderHistory = new ArrayList<>();
+        for (var order : orders) {
+            List<OrderItemsOutput> orderInfo = new ArrayList<>();
+            List<Object[]> orderItemDetails = orderItemsRepo.findByOrderDetailsId(order.orderDetailsId);
+            for (var itemDetail : orderItemDetails) {
+                Product p = (Product) itemDetail[0];
+                OrderItem oi = (OrderItem) itemDetail[1];
+                var orderDetails = new OrderItemsOutput(p.id, p.productName, p.productDescription, p.brand, p.unitPrice, p.unitsInStock, p.size,
+                        p.color, p.productAvailable, p.discontinued, p.discountAvailable, p.picture, p.dateReceived, p.unitsReceived, oi.getQuantity(), oi.getRegularPrice(), oi.getSalePrice());
+                orderInfo.add(orderDetails);
+            }
+            orderHistory.add(new OrderHistoryOutput(orderInfo, order));
+        }
+
+        return orderHistory;
     }
 
     @CrossOrigin
@@ -101,12 +149,76 @@ public class OrderController {
                 Product p = (Product) itemDetail[0];
                 OrderItem oi = (OrderItem) itemDetail[1];
                 var orderDetails = new OrderItemsOutput(p.id, p.productName, p.productDescription, p.brand, p.unitPrice, p.unitsInStock, p.size,
-                        p.color, p.productAvailable, p.discontinued, p.discountAvailable, p.picture, p.dateReceived, p.unitsReceived, oi.getQuantity());
+                        p.color, p.productAvailable, p.discontinued, p.discountAvailable, p.picture, p.dateReceived, p.unitsReceived, oi.getQuantity(), oi.getRegularPrice(), oi.getSalePrice());
                 orderInfo.add(orderDetails);
             }
             orderHistory.add(new OrderHistoryOutput(orderInfo, order));
         }
 
         return orderHistory;
+    }
+
+/*    String createOrder(List<CartItem> cartItems, OrderDetails order) {
+        orderDetailsRepo.save(order);
+        var orderItems = new ArrayList<OrderItem>();
+        var updatedProducts = new ArrayList<Product>();
+        for(var item : cartItems){
+            var prodId = item.getProductId();
+            var itemQty = item.getQuantity();
+            var product = productRepo.findById(prodId).get();
+            //AYE YO FRONTEND!!! WE MIGHT BE ABLE TO CALCULATE CURRENT REGULAR AND SALE PRICES IN THE BACKEND
+            //JAVA HAS METHODS FOR COMPARING DATES
+//            var isBefore = product.scheduledPrices.get(0).getEffectiveDate().before(new Date());
+//            var regPrice = product.scheduledPrices.get(0).getPrice();
+            var unitsInStock = product.unitsInStock;
+            if (itemQty > unitsInStock)
+                return "Sorry, not enough units in stock for " + product.productName + ".";
+            else {
+                var orderItem = new OrderItem(prodId, order.orderDetailsId, itemQty);
+                orderItems.add(orderItem);
+                product.unitsInStock -= itemQty;
+                updatedProducts.add(product);
+            }
+        }
+        productRepo.saveAll(updatedProducts);
+        orderItemsRepo.saveAll(orderItems);
+
+        return "success";
+    }*/
+
+    String createOrder(List<CartItem> cartItems, OrderDetails order, Optional<String> couponCode) {
+        orderDetailsRepo.save(order);
+        var orderItems = new ArrayList<OrderItem>();
+        var updatedProducts = new ArrayList<Product>();
+        for(var item : cartItems){
+            var prodId = item.getProductId();
+            var itemQty = item.getQuantity();
+            var regPrice = item.getRegularPrice();
+            var salePrice = item.getSalePrice();
+            var product = productRepo.findById(prodId).get();
+            //AYE YO FRONTEND!!! WE MIGHT BE ABLE TO CALCULATE CURRENT REGULAR AND SALE PRICES IN THE BACKEND
+            //JAVA HAS METHODS FOR COMPARING DATES
+//            var isBefore = product.scheduledPrices.get(0).getEffectiveDate().before(new Date());
+//            var regPrice = product.scheduledPrices.get(0).getPrice();
+            var unitsInStock = product.unitsInStock;
+            if (itemQty > unitsInStock)
+                return "Sorry, not enough units in stock for " + product.productName + ".";
+            else {
+                var orderItem = new OrderItem(prodId, order.orderDetailsId, itemQty, regPrice, salePrice);
+                orderItems.add(orderItem);
+                product.unitsInStock -= itemQty;
+                updatedProducts.add(product);
+            }
+        }
+        productRepo.saveAll(updatedProducts);
+        orderItemsRepo.saveAll(orderItems);
+
+        if (couponCode.isPresent()) {
+            var coupon = couponRepo.findById(couponCode.get()).get();
+            coupon.orders.add(order);
+            couponRepo.save(coupon);
+        }
+
+        return "success";
     }
 }
